Перенести тесты в абстрактный класс TaskManagerTest
